requirements.txt
groq
sounddevice
soundfile
numpy
keyboard
pyperclip
python-dotenv
pynput
Pillow
PySide6
pyinstaller
